Add example of address pattern to error summary
If you use the error summary component with our current address pattern (multiple text inputs within a fieldset), clicking on the error message for the post code scrolls to the top of the fieldset which means e.g. on iOS the postcode field ends up obscured by the keyboard.

This example demonstrates that and allows us to test for sensible behaviour.