Fix text resize issue with warning text icon
The font size of the exclamation mark within the warning text component is currently specified as 1.6 em, within a 19px context which means that the calculated font-size is 30.4px.

Because the height and width of the circle the exclamation mark sits within are specified in px, the exclamation mark increases in size when the page is zoomed text-only, whilst the circle does not. This results in the mark overflowing outside of the circle.

Specifying the font-size in px solves this problem. Whilst the icon is used to add emphasis, the text is also inset and bold, so we can treat it as presentational.