Remove redundant code, move code for moving element in DOM
This removes the branching logic and the code that added the count message markup dynamically in `createCountMessage()`.
It’s worth nothing that the code inside the block hasn’t been executed since https://github.com/alphagov/govuk-frontend/commit/7b534bf734cd0a1be2cff34d64018aa89abdabb6.
As using the character count with the count message markup removed has never been part of the public API, removing this functionality
is not considered a breaking change.

The original code for checking/creating the count message markup [moves it in the DOM](https://github.com/alphagov/govuk-frontend/blob/f4818118a71462972d4ec5d4b6409e89c8fbc56c/src/govuk/components/character-count/character-count.js#L104) to be inside
`govuk-form-group`. This was probably useful when the message markup was created dynamically. This PR retains the bit of code that
moves the element for backwards compatibility. It’s important that the count message markup is inside `govuk-form-group` so that it gets
wrapped within the error block. However we should look to remove moving it with JS and instead change the component markup so that
the count message markup is inside `govuk-form-group`. This would be a breaking change so shall leave it for now and create an issue.