Ensure footer links look clickable
We originally removed the underlines from these links to make them consistent with the approach we have for the header component.

In order to make sure these links look like links in this component I have reverted this change.

Original reasoning for removing underlines:
"
I believe the intention was to treat the links in the footer in the same way we treat other navigational elements – like the header, or in the case of the Design System the side navigation. Where the ‘thing’ is a ‘list of links’, the thinking is/was that they don’t need the underlining because there’s no wider context in which they need to be distinguished. The exception is any link that does appear within e.g. a paragraph, in which case it does need the underline to distinguish it. It has been questioned a few times on support recently though.

The benefit, I believe, is that it allows us to have a clear hover state for those links, which we didn’t previously have
"