Fix warning text if border-box sizing applied globally
If a user applies a global box-sizing reset [1], the warning text icon is 'squished' horizontally as the border is 'accounted for' within the 29px min-width.

Explicitly use border-box sizing, as this means that the min-width reflects the actual desired width of the icon, rather than having to factor the border width in separately. As the box-sizing model is now set explicitly, it should display correctly whether the reset is applied or not.

Increase the min-width and min-height of the icon to include the border (3px border * 2 in each direction = 6px).

[1]: https://css-tricks.com/inheriting-box-sizing-probably-slightly-better-best-practice/