Autofocus notification banner when the page loads
If a notification banner has the `data-initial-focus` focus attribute,
it should be focused on page load to ensure that it is announced to the user
(in addition, the component HTML sets `role="alert"` for increased visibility
to assistive technologies).

Scroll the element to the viewport when focused.

Don't focus the element if there is an error summary on the page, or another
notification banner has already received focus.