Simplify back link styles
We used to use a bottom border to create the underline so that the ‘triangle’ and the link were both underlined.

This was changed in ad141f4 to make it consistent with breadcrumbs, which means the special link treatment isn’t required any more.

Remove it, as it’s preventing the back link from inheriting the new link decoration styles.